requirements.txt
# API and frontend
fastapi==0.110.0
uvicorn[standard]==0.29.0
streamlit==1.31.0
requests==2.32.3
python-multipart==0.0.20
python-dotenv==1.0.1

# RAG pipeline
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7
langchain-text-splitters==0.3.2
langchain-google-genai==2.0.4
google-generativeai==0.8.3
langchain-qdrant==0.2.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
pypdf==5.1.0

# Data & Charts (dashboard)
pandas>=2.0.0
plotly>=5.18.0